refactor: 역 관리 기능 호출 구조 변경
